chore: allow any CORS origin by default
