+++ Desktop/KurierStrona/.gitignore
