+++ Desktop/KurierStrona/frontend/src/api.ts
@@ -1,0 +1,16 @@
+import axios from 'axios'
+
+const api = axios.create({
+  baseURL: import.meta.env.VITE_API_URL || '/api',
+})
+
+api.interceptors.request.use((config) => {
+  const token = localStorage.getItem('token')
+  if (token) {
+    if (!config.headers) config.headers = {} as any
+    ;(config.headers as any).Authorization = `Bearer ${token}`
+  }
+  return config
+})
+
+export default api
